Add gotcha about accessing non-existent properties
JavaScript does not issue a warning, etc. when you access a non-existent
property. It simply returns `undefined`. Moreover, JavaScript treats
`undefined` just like another value, and it allows comparisons on it
without throwing an exception, issuing a warning, etc. This may cause
subtle bugs. The solution is to use TypeScript (or a static type
checking library).