Later logs of an error are identical to the first
When you log an error using `console.log` once, the subsequent logs of
the error will be identical to the first log's output, even if the
`message` property of the error is changed between the first and the
second log. Note that second log might be `console.log`, OR it might be
the print-out of the error being thrown and uncaught.